merging

+++ Frontend/src/pages/Category.jsx
@@ -1,10 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect,useContext } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
 import HorizontalCardScroll from '../Components/HorizontalCardScroll';
 import axios from "axios";
 import ReviewSection from '../Components/ReviewSection';
 import { BsCart3 } from "react-icons/bs";
 import Swal from 'sweetalert2';
+import { AuthContext } from '../Components/AuthProvider'; 
 import { toast } from 'react-toastify';
 
 const Category = () => {
@@ -17,6 +18,7 @@ const Category = () => {
   const [items, setItems] = useState([]);
   const [cartItems, setCartItems] = useState([]);
   const [itemAddedToCart, setItemAddedToCart] = useState(false);
+  const userId = localStorage.getItem('userId');
 
   useEffect(() => {
     const fetchData = async () => {
@@ -33,12 +35,14 @@ const Category = () => {
     };
 
     // Retrieve existing cart items from local storage once during initial mount
+    fetchData();
+  }, []);
+
+  useEffect(() => {
     const storedCartItems = localStorage.getItem('cartItems');
     if (storedCartItems) {
       setCartItems(JSON.parse(storedCartItems));
     }
-
-    fetchData();
   }, []);
 
   if (!product) {
@@ -55,18 +59,37 @@ const Category = () => {
     setQuantity(Math.max(1, value));
   };
 
+  const { isAuthenticated ,user} = useContext(AuthContext);
+
+  
   const handleAddToCart = (product, quantity) => {
+
+  if (!isAuthenticated) {
+    Swal.fire({
+      title: "Please Login",
+      timer: 2000, // Close the alert after 2 seconds
+      icon: "warning",
+      timerProgressBar: true,
+      showConfirmButton: false // Hide the OK button
+    }).then(() => {
+      navigate("/login"); // Navigate to cart after alert is closed
+    });
+    // Handle case when user is not logged in
+    console.log('User is not logged in. Please log in to add items to the cart.');
+    return;
+  }
+
     const cartItem = {
       ...product,
       quantity,
       isMedicine: !!product.Medicine_Name,
+      userId:user.id,
     };
-    setCartItems((prevCartItems) => {
-      const updatedCartItems = [...prevCartItems, cartItem];
-      localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
-      setItemAddedToCart(true); // Set the state to indicate item added to cart
-      return updatedCartItems;
-    });
+    const updatedCartItems = [...cartItems, cartItem];
+    setCartItems(updatedCartItems);
+    localStorage.setItem('cartItems', JSON.stringify(updatedCartItems));
+    setItemAddedToCart(true); // Set the state to indicate item added to cart
+    console.log('Product added to cart:', cartItem);
   };
 
   // If item added to cart, display success message and navigate to cart
